+++ src/invoices/cc-invoice-list.js
@@ -119,12 +119,14 @@ export class CcInvoiceList extends withResizeObserver(LitElement) {
                   legend=${i18n('cc-invoice-list.year')}
                   .choices=${yearChoices}
                   value=${yearFilter}
+                  inline
                   @cc-toggle:input=${this._onYearFilterValue}
                 ></cc-toggle>
                 <cc-select
                   label=${i18n('cc-invoice-list.year')}
                   .options=${yearChoices}
                   value=${yearFilter}
+                  inline
                   @cc-select:input=${this._onYearFilterValue}
                 ></cc-select>
               ` : ''}